Separate out commands and add git primitives for fetching and merging
